08.	Company Users
Write a program that keeps information about companies and their employees. 
You will be receiving a company name and an employee's id, until you receive the command "End" command. Add each employee to the given company. Keep in mind that a company cannot have two employees with the same id.
When you finish reading the data, order the companies by the name in ascending order. 
Print the company name and each employee's id in the following format:
{companyName}
-- {id1}
-- {id2}
-- {idN}
Input / Constraints
•	Until you receive the "End" command, you will be receiving input in the format: "{companyName} -> {employeeId}".
•	The input always will be valid.